gcp/requirements.txt
google-cloud-storage==2.4.*
scikit-learn==1.4.0
h5py==3.10.0
flask==2.3.3
flask-cors==4.0.0
numpy==1.26.3